Implemented the QA displaying feature
The questions and their corresponding answers within this list will be displayed in an expanding and collapsing accordion. By default, on page load up to four questions should be displayed. Up to two answers should display for each question. The remaining questions or answers should be hidden until the user loads them using the “More Answered Questions” button (section 1.3.4).

+ Added a personal utility file to do customized tasks:
a. turn an object of objects (where objects are similar) to an array of objects
b. topXItems to create an array that has a chosen number of items from an array

+ Installed Underscore