1.3.2.2 Individual Questions - Helpful update
A link should appear next to the text “Helpful?” reading “Yes (#)” with the count of selections for that answer. Clicking on this link should increase the count for that response. A customer should not be able to vote more than once for this selection.

Add Axios (axios.create) to the utility file