Update Question.jsx
Next to the link for “Helpful?”, a second link reading “Report” will appear. Clicking on this link will mark the answer for internal review. A user should not be able to report an answer more than once. After clicking on this link, the “Report” link should change to static text that reads “Reported”. Answers that have been reported should be marked as such in the system for further action to be taken.